add stub 64bit

+++ xposedcompat/src/main/java/com/swift/sandhook/xposedcompat/hookstub/CallOriginCallBack.java
@@ -1,5 +1,5 @@
 package com.swift.sandhook.xposedcompat.hookstub;
 
 public interface CallOriginCallBack {
-    int call(long... args) throws Throwable;
+    long call(long... args) throws Throwable;
 }
